This is our smart radar's code and smart watch code.
In this code we use a naive learning method to learn the enviorment's characteristic distance ,to save arduino's dynamic memory we use dynamic memory allocation to create and reuse the array to collect statistics data. Also, we use simple average method and standard deviation conception to filter the data and increase its accuracy.
In smart watch code we use many interrupt to deal with multitask problem.